Fix IP issue and repo reorganization
The IP issue that would cause "localhost" to appear might be fixed.
It seems the error became permanent with Android 13, and now appears to
be fixed.

Now the repo has all the files directly in it and the ps3proxy Go file
is now a (small) project with go.mod included.

Also updated the ps3proxy Go file to try/catch any fatal error that
could appear for any reason (just in case).

The app was updated to use Android X too and is compiling with Android
Studio normally right now.

I didn't improve anything else on the app. It needs a whole rewrite...
I wish I could do it but I needed more time. Some day when no one needs
it anymore I'll update it xD. Was my first app. The code is a mess...
But I got it to work back then at least xD.